chore: split up GhaFS into separate file
